requirements-bkp.txt
absl-py==0.15.0
albumentations==1.3.0
asttokens==2.1.0
astunparse==1.6.3
backcall==0.2.0
batchgenerators==0.24
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
flatbuffers==1.12
fonttools==4.38.0
future==0.18.2
gast==0.3.3
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.4
imageio==2.22.4
importlib-metadata==5.0.0
ipykernel==6.17.1
ipython==8.6.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==7.4.7
jupyter_core==5.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
linecache2==1.0.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
monai==1.0.1
mpi4py==3.0.3
nest-asyncio==1.5.6
networkx==2.8.8
numpy==1.19.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
prompt-toolkit==3.0.33
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qudida==0.0.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
SimpleITK==2.2.0
six==1.15.0
stack-data==0.6.1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.10.10
torch==1.13.0
tornado==6.2
tqdm==4.64.1
traceback2==1.4.0
traitlets==5.5.0
typing-extensions==3.7.4.3
unittest2==1.1.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.12.1
zipp==3.10.0
